# docs/api/openapi-mvp.yaml
openapi: 3.0.3
info:
  title: Moimday API
  version: 1.3.0
  description: MVP REST API ??SSOT with api-contract-mvp.md (multi-group, 30 members)
servers:
  - url: http://localhost:3000/v1
    description: Local dev
  - url: https://api.Moimday.example/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /events/{eventId}:
    get:
      tags: [events]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Event detail }
    patch:
      tags: [events]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Updated }
        '409': { description: VERSION_MISMATCH }
  /events/{eventId}/date-votes:
    put:
      tags: [events]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Votes saved }
  /events/{eventId}/responses:
    put:
      tags: [events]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Response saved }
  /events/{eventId}/confirm-datetime:
    post:
      tags: [events]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: attendance_open }
  /events/{eventId}/finalize:
    post:
      tags: [events]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: finalized }
  /events/{eventId}/cancel:
    post:
      tags: [events]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: cancelled }
  /events/{eventId}/nudge:
    post:
      tags: [events]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Nudge sent }
  /events/{eventId}/date-poll-summary:
    get:
      tags: [events]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Poll summary per option (up to 30 members per bucket)
  /events/{eventId}/comments:
    get:
      tags: [events]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Comments }
    post:
      tags: [events]
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Comment created }
  /comments/{commentId}:
    delete:
      tags: [events]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: commentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Deleted (author, organizer, group admin) }
        '403': { description: FORBIDDEN }
  /events/{eventId}/extend-poll-deadline:
    post:
      tags: [events]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pollDeadlineAt: { type: string, format: date-time }
                addOptions:
                  type: array
                  items:
                    type: object
                    properties:
                      startsAt: { type: string, format: date-time }
                      isAllDay: { type: boolean }
      responses:
        '200': { description: Extended }
  /events/{eventId}/date-options:
    put:
      tags: [events]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Options replaced; votes reset }
...